# Optimize the kernel implementation of layernorm with openmp