@luotao1 2019-10-31
Ernie性能数据在：https://github.com/PaddlePaddle/benchmark/issues/180#issuecomment-548204570 @Xreki 可以再看一下